feat: Introduce app
