+++ app/global/constants/resources/images/index.ts
@@ -3,6 +3,15 @@ export const ImagesConstants = {
   icons: {
     google: require("@/../assets/images/icons/search.png"),
     facebook: require("@/../assets/images/icons/facebook.png"),
-    intro1: require("@/../assets/images/app/intro.png"),
   },
+  backgrounds: {
+    intro: {
+      slide1: require("@/../assets/images/backgrounds/bg-introstep1.png"),
+      slide2: require("@/../assets/images/backgrounds/bg-introstep2.png"),
+      slide3: require("@/../assets/images/backgrounds/bg-introstep3.png"),
+    }
+  },
+  button: {
+    iconArrowRight: require("@/../assets/images/buttons/btn-arrow-right.png"),
+  }
 };
