+++ package.json
@@ -3,6 +3,7 @@
     "@babel/eslint-parser": "^7.23.10",
     "@babel/plugin-proposal-class-properties": "^7.18.6",
     "@babel/plugin-proposal-decorators": "^7.23.9",
+    "@babel/plugin-proposal-export-namespace-from": "^7.18.9",
     "@devvie/bottom-sheet": "^0.3.0",
     "@expo-google-fonts/inter": "^0.2.3",
     "@expo-google-fonts/rowdies": "^0.2.3",
@@ -17,23 +18,23 @@
     "bloc-builder-react": "^1.0.1",
     "eslint": "^8.56.0",
     "eslint-config-universe": "^12.0.0",
-    "expo": "~50.0.5",
-    "expo-blur": "~12.9.1",
+    "expo": "~50.0.7",
+    "expo-blur": "~12.9.2",
     "expo-constants": "~15.4.5",
     "expo-font": "~11.10.3",
-    "expo-image": "~1.10.5",
-    "expo-linear-gradient": "~12.7.1",
+    "expo-image": "~1.10.6",
+    "expo-linear-gradient": "~12.7.2",
     "expo-linking": "~6.2.2",
     "expo-localization": "~14.8.3",
-    "expo-router": "~3.4.6",
+    "expo-router": "~3.4.7",
     "expo-splash-screen": "~0.26.4",
     "expo-status-bar": "~1.11.1",
     "i18n-js": "^4.3.2",
     "prettier": "^3.2.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-native": "0.73.2",
-    "react-native-gesture-handler": "^2.15.0",
+    "react-native": "0.73.4",
+    "react-native-gesture-handler": "~2.14.0",
     "react-native-reanimated": "~3.6.2",
     "react-native-safe-area-context": "4.8.2",
     "react-native-screens": "~3.29.0",
